// DVG.CRM.XeCung.ClientSide/dvg-crm-xecung-final-client/src/app/models/usersearchresult.ts
export class UserSearchResult {
    constructor() {}
    Index: number;
    Id: number;
    UserName: string;
    Email: string;
    DisplayName: string;
    CreatedDate: string;
    CreatedBy: string;
    Status: number;
    LastActivityDate: string;
    PhoneNumber: string;
    Position: string;
    StrLastActivityDate: string;
}