// DVG.CRM.XeCung.ClientSide/dvg-crm-xecung-final-client/src/app/models/lead-search.ts
export class LeadSearch {
    constructor() {}
    Id: number;
    AdminId: number;
    PhoneNumber: string;
    Email: string;
    Address: string;
    CrawleredPostingNews: number;
    FullName: string;
    Note: string;
    AssigneeName: string;
    AssigneeId: number;
    IsCheckedAssign: boolean;
    BlockStatus: number;
    BlockReasonDescription: string;
    BlockReasonType: number;
}